+++ lib/views/screens/barcode_scanner_page.dart
